Title: 소수의 거듭제곱, Time: 247ms, Memory: 16MB, Status: Wrong Answer - CodeTree

+++ 240729/소수의 거듭제곱/power-of-prime-number.java
@@ -26,7 +26,7 @@ public class Main {
             if (!isPrime[i]) {
                 continue;
             }
-            long j = i * i;
+            long j = (long)i * i;
             while (j <= b) {
                 if (a <= j) {
                     answer++;
